// lib/types/models/project.ts
export type ResourceType =
  | 'user-files'
  | 'template-collections'
  | 'github-installations'
  | 'chat-conversations'
  | 'project-chat-conversations'
  | 'project-context-files'
  | 'project-context-imports'
  | 'project-resources'
  | 'project-template-collections'
  | 'project-user-files'
  | 'project-github-installations'
  | 'project-context-imports'
  | 'project-context-files'
  | 'project-chat-conversations'